@@ -7,23 +7,23 @@
   "license": "GPL-3.0",
   "dependencies": {
     "@cypress/react": "^5.12.4",
-    "bulma": "^0.9.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^4.0.3"
+    "@fortawesome/fontawesome-free": "^6.2.0",
+    "bulma": "^0.9.4",
+    "classnames": "^2.3.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1"
   },
   "devDependencies": {
     "@cypress/webpack-dev-server": "^1.8.4",
-    "@mate-academy/cypress-tools": "^1.0.4",
-    "@mate-academy/eslint-config-react": "*",
-    "@mate-academy/eslint-config-react-typescript": "*",
-    "@mate-academy/scripts": "^1.2.8",
+    "@mate-academy/cypress-tools": "^1.0.5",
+    "@mate-academy/eslint-config-react-typescript": "^1.0.12",
+    "@mate-academy/scripts": "^1.2.12",
     "@mate-academy/students-ts-config": "*",
-    "@mate-academy/stylelint-config": "*",
-    "@types/node": "^17.0.23",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
-    "cross-env": "^7.0.3",
+    "@mate-academy/stylelint-config": "^0.0.11",
+    "@types/node": "^17.0.45",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "cypress": "^9.5.3",
     "eslint": "^7.32.0",
     "eslint-plugin-cypress": "^2.11.2",
@@ -31,8 +31,9 @@
     "mochawesome": "^7.1.3",
     "mochawesome-merge": "^4.2.0",
     "mochawesome-report-generator": "^6.2.0",
-    "node-sass": "^6.0.1",
-    "stylelint": "^13.13.1",
+    "node-sass": "^7.0.1",
+    "postcss": "^8.4.12",
+    "stylelint": "^15.11.0",
     "typescript": "^4.6.3"
   },
   "scripts": {
