@@ -1,6 +1,6 @@
-import ReactDOM from 'react-dom';
 import 'bulma/css/bulma.css';
-
+import { createRoot } from 'react-dom/client';
 import { App } from './App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+createRoot(document.getElementById('root') as HTMLElement)
+  .render(<App />);
